♻️ refactor(hotkeys): consolidate desktop global shortcut definitions (LOBE-7181)
Made-with: Cursor
