+++ src/hooks/useHotkeys/useHotkeyById.test.ts
@@ -1,9 +1,9 @@
+import { HotkeyEnum, HotkeyScopeEnum } from '@lobechat/const/hotkeys';
 import { renderHook } from '@testing-library/react';
 import { uniq } from 'es-toolkit/compat';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { HOTKEYS_REGISTRATION } from '@/const/hotkeys';
-import { HotkeyEnum, HotkeyScopeEnum } from '@/types/hotkey';
 
 import { useHotkeyById } from './useHotkeyById';
