+++ src/routes/(main)/agent/features/Conversation/MainChatInput/useSendMenuItems.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { HotkeyEnum, KeyEnum } from '@lobechat/const/hotkeys';
 import { type MenuProps } from '@lobehub/ui';
 import { Flexbox, Hotkey, Icon } from '@lobehub/ui';
 import { BotMessageSquare, LucideCheck, MessageSquarePlus } from 'lucide-react';
@@ -10,7 +11,6 @@ import { useConversationStore, useConversationStoreApi } from '@/features/Conver
 import { useAddUserMessageHotkey } from '@/hooks/useHotkeys';
 import { useUserStore } from '@/store/user';
 import { preferenceSelectors, settingsSelectors } from '@/store/user/selectors';
-import { HotkeyEnum, KeyEnum } from '@/types/hotkey';
 
 /**
  * useSendMenuItems hook for ConversationStore
